Decouple inspect data grabbing from TreeViewProvider
